requirements.txt
asgiref==3.4.1
coverage==6.2
Django==4.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.2
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
